Refactor PROB_LIST macros
We're never using PROB_LIST_GET without "random" as the item. It looks
like it's original intention was to aid in modifying prob_lists, which I
would like to change so it never happens. I would like to eventually
replace all code that modifies prob_lists to one that uses the new
late-expansion functionality.