GAME: revert vghetto changes to dark magic requirements
This is fairly complex and it appears that there used to be two
different ways a human could end up with the dark magic path.
Historically, if one was already a human magic initiate, it cost zero
points to "go to the dark side". This reversion sets that back to the
case and ostensibly disables the "dark magic initiate" path, meaning you
will have to first choose "magic initiate" and then "dark magic" for
zero talent points.

There's a WHOLE LOT of fixing to be done on dark magic, as soultrap and
plague appear to be non-functioning at the moment and after reading the
history in the Wesband thread, they may have been over-powered as well.